app/requirements.txt
aiofiles==23.1.0
aioredis==2.0.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
ecdsa==0.18.0
fastapi==0.94.1
h11==0.14.0
idna==3.4
passlib==1.7.4
Pillow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.6
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
text-unidecode==1.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
